feat: Upgrade black, click and gitlint versions (more info below)
For some reason after making the change to the yaml.serialize to
include the encoding keyword explicitly, black started failing with
the message ImportError: cannot import name '_unicodefun' from 'click'.
Found this solution in SO: https://stackoverflow.com/a/71674345/3716354
which involved upgrading black. However, in order to upgrade it I would
need to upgrade click, which would need to upgrade gitlint. After upgrading
black, it required pre-commit to be upgraded as well. So that's
why we have more package upgrades in the following order:

gitlint > click > black > pre-commit