feat: upgrade Python version and some of its packages (#549)
* BREAKING CHANGE: Upgrade the minimal Python version in pyproject.toml

The end of life for the Python 3.6 was in december 2021, so I think
it's the right time to drop support for this version

* feat: Upgrade PyYAML and types-PyYAML packages

* feat: Upgrade Jinja to the latest version

* chore: Updates changelog with the appropriate changes

* fix: Upgrade Python version in github workflows

* feat: Upgrade black, click and gitlint versions (more info below)

For some reason after making the change to the yaml.serialize to
include the encoding keyword explicitly, black started failing with
the message ImportError: cannot import name '_unicodefun' from 'click'.
Found this solution in SO: https://stackoverflow.com/a/71674345/3716354
which involved upgrading black. However, in order to upgrade it I would
need to upgrade click, which would need to upgrade gitlint. After upgrading
black, it required pre-commit to be upgraded as well. So that's
why we have more package upgrades in the following order:

gitlint > click > black > pre-commit

* Update black rev in .pre-commit-config.yaml

* fix: pass encoding keyword explicitly in yaml.serialize function

* fix: Upgrade Poetry version to 1.1.3 to fix maximum recursion error

* fix: Upgrade Poetry version to fix recursion error

* fix: Upgrade Poetry version in GitHub action files (now for good)

* chore: Run black to reformat some files

* chore: Upgrade mypy to the latest version

* fix: Remove encoding from yaml.serialize call

* chore: Update CHANGELOG.md

Co-authored-by: Camila Maia <cmaiacd@gmail.com>

Co-authored-by: Camila Maia <cmaiacd@gmail.com>